% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_cox}
\alias{build_cox}
\title{Builds a function used to calculate hazard ratios.}
\usage{
build_cox(outcome, covars = NULL)
}
\arguments{
\item{var}{The dependent variable}

\item{data}{The data frame.}
}
\value{
A vector of strings with the hazard ratios.
}
\description{
Builds a function to build a Cox model on the
outcome survival object, with a single covariate and
returns a list of the odds ratios for each level of the factor.
}
\keyword{internal}
