% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table.R
\name{convert_list_to_matrix}
\alias{convert_list_to_matrix}
\title{Converts a table with summary values saved as a list to a matrix.}
\usage{
convert_list_to_matrix(x)
}
\arguments{
\item{list}{Input list}
}
\value{
A string matrix containing the content of each cell

Internal helper function
}
\description{
Converts a table with summary values saved as a list to a matrix.
}
\keyword{internal}
