% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_functions.R
\name{freq}
\alias{freq}
\title{Calculates cross-tabulated frequencies}
\usage{
freq(proportion = c("percentage", "ratio", "none"), digits = 2)
}
\arguments{
\item{proportion}{Whether to include the proportion of counts this represents in parantheses.}

\item{digits}{The number of digits to specify \code{proportion} to.}
}
\value{
A function that calculates the frequency of a cell in a contingency table.
}
\description{
Calculates cross-tabulated frequencies
}
