% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neat_table.R
\name{neat_table}
\alias{neat_table}
\title{Displays a neatly formatted table.}
\usage{
neat_table(table, format = "html")
}
\arguments{
\item{table}{A \code{crosstab} object.}
}
\value{
A \code{kable} object.
}
\description{
Displays a neatly formatted table.
}
