% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_contintab.R
\name{print.contintab}
\alias{print.contintab}
\title{Prints the contingency table as an ASCII table}
\usage{
\method{print}{contintab}(x, ...)
}
\arguments{
\item{x}{An object of class \code{contintab} return by
\code{contingency_table}.}

\item{...}{Options passed to \code{print}.}
}
\value{
None. Prints a table to standard output.
}
\description{
Prints the contingency table as an ASCII table
}
