% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_funtions.R
\name{summary_mean}
\alias{summary_mean}
\title{Calculates the mean of a specified continuous variable
for each level of an outcome.}
\usage{
summary_mean(var)
}
\arguments{
\item{var}{A continuous variable name as a string.}
}
\value{
A function that calculates the mean value of \code{var}
for each outcome level.
}
\description{
Calculates the mean of a specified continuous variable
for each level of an outcome.
}
