% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_funtions.R
\name{summary_median}
\alias{summary_median}
\title{Builds a function to calculate the median of a
continuous variable for each level of an outcome.}
\usage{
summary_median(var)
}
\arguments{
\item{var}{A continuous variable name as a string.}
}
\value{
A function that calculates the median value of \code{var}
for each outcome level.
}
\description{
Builds a function to calculate the median of a
continuous variable for each level of an outcome.
}
