% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_functions.R
\name{freq}
\alias{freq}
\title{Builds a function to calculate cross-tabulated frequencies}
\usage{
freq(proportion = c("percentage", "ratio", "none"), digits = 2)
}
\arguments{
\item{proportion}{If 'percentage' or 'ratio', the proportion of counts
is included in parentheses.}

\item{digits}{The number of digits to specify \code{proportion} to.}
}
\value{
A function that calculates the frequency of a cell in a contingency table.
}
\description{
Builds a function to calculate cross-tabulated frequencies
}
