% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neat_table.R
\name{neat_table}
\alias{neat_table}
\title{Displays a neatly formatted table.}
\usage{
neat_table(table, format = c("html", "pdf"), ...)
}
\arguments{
\item{table}{A \code{contintab} object.}

\item{format}{A string specifying output format passed to \code{knitr::kable}.
Currently only 'html' and 'pdf' are supported.}

\item{...}{Other arguments passed to \code{knitr::kable}.}
}
\value{
A \code{kable} object.
}
\description{
Displays a neatly formatted table.
}
