% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Builds a function used to calculate odds ratios.}
\usage{
odds_ratio(outcome, covars = NULL)
}
\arguments{
\item{var}{The dependent variable}

\item{data}{The data frame.}
}
\value{
A vector of strings with the hazard ratios.
}
\description{
Builds a function to calculate logistic regression on the
outcome variable of interest with a single covariate and
returns a list of the odds ratios for each level of the factor.
}
\keyword{internal}
