% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Builds a function used to calculate odds ratios.}
\usage{
odds_ratio(outcome = NULL, adjusted = FALSE, relevel_baseline = TRUE)
}
\arguments{
\item{outcome}{The dependent variable, defaults to the first \code{strata}
supplied to \code{contingency_table}.}

\item{adjusted}{Whether to adjust for the other covariates, specified by
\code{cat_vars} argument to \code{contingency_table}.}

\item{relevel_baseline}{Whether to use the largest level as the baseline.}
}
\value{
A function that is used to calculate odds ratios.
}
\description{
Builds a function to calculate logistic regression on the
outcome variable of interest with a varying number of covariates and
returns a list of the odds ratios for each level of the factor.
}
